Add Skip process and CSP enum type
We need the CSP enum type now to be able to test the Skip process, since
its outgoing transitions mention Stop.  That means we need an enum of
Skip and Stop, and since those are the only two processes we've defined
at this point, we can go ahead and call that enum "CSP"!  We'll add more
to it as we add more primitive processes and operators.