Generate Tau and Tick for arbitrary TestEvents
Before we left those out so that we wouldn't name them directly when
constructing prefix processes.  We still want to prevent that, but this
does so _only_ for prefix processes.  For all other cases when you ask
for an arbitrary TestEvent, you can now get Tau and Tick.  We'll be
using TestEvent parameters more often to test the outgoing transitions
of a process — where we do want to verify things about Tau and Tick —
than using them to construct arbitrary processes.  (For that we'll
usually just ask for a CSP<TestEvent> directly, which knows how to do
the right thing for prefix processes.)