Add lifetimes to initials and afters results
This lets us return iterators that have references back to the process
itself.  The values coming _out_ of the iterator still need to be
cloned, since the `Item` type is `P`, not `&'a P`.  (We can't change the
iterators to return a reference directly, since some operators will need
to construct their after processes on the fly.  We can _probably_ have
the output be a Cow, but I'll save that for a separate patch.)