Simplify afters result
We don't need to wrap the result in an Option; we can use an empty
result iterator to indicate that the process doesn't respond to a
particular event.  I had originally gone with an Option wrapper in an
attempt to make it easier to implement the afters methods — in many
cases, you could return an existing iterator wrapped in Some, without
having to create a custom type for the return value.  But almost all of
the operators are going to need custom return types _anyway_, so we're
not really saving any effort there.