Break recursion in CSP type
We'll need this to implement `initials` or `afters` for any operator
that needs to recurse into its subprocesses.  Without it, we'd need to
prove that CSP implements Initials in order to prove that (for instance)
ExternalChoice does; but we'd also need to prove that ExternalChoice
implements Initials to prove that CSP does!  Whoops.