Add Process and Cursor traits
I want to try out a different approach to defining a process's
semantics.  Instead of return the "afters" as another _process_, instead
each Process implementor will define a separate Cursor type, which tracks
the current state of the process as events occur.  The Cursor type,
importantly, would correspond to a _set_ of after processes — kind of
like a prenormalized process, but with tau events allowed.  Let's see
how it goes!