Add Possibilities type and perform_piecewise
A lot of CSP operators have semantics that will need to keep track of a
set of possible states that the process could be in, where each state is
somehow derived from the states that its subprocesses are in.  And there
is (I think, blindly) a small set of common operations that will be
performed on those sets of possibilities.  So this patch adds a new
helper type so that we don't have to go off and implement this pattern
multiple times.  The first common operation is `perform_piecewise`,
which allows each subprocess to perform an event "independently".  If
more than one subprocess in a possible state can perform the event, then
you end up with additional possible states afterwards — in each one,
exactly one of the subprocesses is allowed to perform the event, and the
others stay in their before state.