Add perform_all method
This is another pattern for dealing with possible sets of subprocess
states.  It has every subprocess perform an event _if it can_.  This is
not parallel-style synchronization — if some subprocesses can't perform
the event, that doesn't prevent the ones that can from doing so.  This
will be used, for instance, for the τ event of an internal choice, and
for any follow-on events of an internal or external choice.