Track multiple Q states in sequential choice
We can shift over from P to Q in a sequential choice at multiple points
in a trace, depending on exactly where P can perform τ and ✔ events.
But we were only keeping track of a single current state for the Q
process.  Instead we need to keep track of the current state of Q for
all of the possible places earlier in the trace where we might have
shifted over to Q.