Add Alphabet trait
This abstracts away the idea of an _alphabet_, which is a set of events.
Before, we were using Iterators and HashSets for this, but we don't want
to assume that events are always easy to enumerate and store into an
actual set.  You might want to define them based on a predicate, for
instance!  This trait will hopefully make it possible to do that.