Clear P side of seq-comp after performing ✔
We were correctly clearing the P side of a sequential composition when
we were trying to perform a visible event that P couldn't perform — that
indicates that the visible event is coming from one of our Q processes,
and P is no longer a valid possibility.  But we weren't doing that when
we'd try to perform a ✔.  The P side of a sequential composition can
_never_ perform a ✔, and so we should always clear the P side when that
event occurs — it can only be coming from the Q side.