Add EventSet trait
This is the replacement for the old Alphabet trait.  Alphabet was
generic over a type representing individual events.  EventSet doesn't
have this — individual events are instead just represented as a set
containing a single element.

With Alphabet, we would handle different "kinds" of event by creating a
sum type to use as that generic event type.  Each variant of the enum
would handle each of the different kinds of event.  With EventSet, each
kind of event will be represented by a separate type, and it is the
_set_ types that will be enum-ed together, instead of the _event_ types.

Each EventSet implementation needs to be able to create `empty` and
`universe` sets, which respectively contain no elements and every
possible element.

There's also no `contains` method in EventSet (since you can't construct
an individual event to check for).  Instead there are several methods
for typical set operations (union, intersection, difference, negation).