Add maximal_finite_traces for new CSP type
This is exactly like the existing maximal_finite_traces function, but it
now works with the new EventSet trait and CSP type.  As with the other
patches in this series, the resulting maximal finite traces is a list of
singleton event sets, since we don't have types for individual events.