Add Tau and Tick traits
With the new EventSet trait, we don't represent τ and ✔ as distinct
values.  Instead, we create a trait for each of them, that can check
whether a particular EventSet contains τ or ✔.  If it's possible for an
EventSet implementation to contain these events, that type will _also_
implement Tau and Tick.  Process operator implementations will be able
to include those in their constraints, letting them check whether
particular event sets contain τ and ✔.

We also create a new PrimitiveEvents type, which implements EventSet,
and whose universe just consists of τ and ✔.  We can't yet combine that
into larger event sets, containing other kinds of events, but we will
soon.