Add DisjointSum type
This type is how we combine two different "kinds" of event into a type
that can hold both.  It relies on the essential assumption that the two
kinds of event are disjoint: there is no event that is a valid instance
of both kinds.  This is very similar to how events are handled in CSPM:
the `channel` statement introduces a completely new set of events that
is entirely disjoint from any other events that exist in the model.

With this representation, we can implement all of the set operations by
delegating them to the underlying "variants" of the disjoint sum —
meaning that a DisjointSet automatically implements EventSet if both of
its variants do.

The only wrinkle is that we have to modify our traits for tracking
whether a particular EventSet type can contain τ and ✔.  They now have
to carry "proof" in the form of a generic type parameter.  This borrows
a trick from the frunk crate [1], which lets us implement Tau and Tick
for both sides of DisjointSum without introducing overlapping impls.

[1] https://beachape.com/blog/2017/03/12/gentle-intro-to-type-level-recursion-in-Rust-from-zero-to-frunk-hlist-sculpting/