Add iterators for each event set type
Not every event set type will support this, but all of the ones we've
defined so far will.  And we'll need to be able to enumerate all of the
events in a set to calculate a process's maximal finite traces, which
will be an important part of our test cases.

Note that unlike with Alphabet, we don't have a concrete type to hold
individual events, so these iterators enumerate through all of the
one-element subsets of the event set.